backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
bcrypt
PyJWT
pandas
openpyxl
playwright
python-multipart
python-dotenv
requests
